cppcheckexecutor.cpp: extracted `USE_UNIX_SIGNAL_HANDLING` code to separate file
