+++ cli/cli.vcxproj.filters
@@ -29,6 +29,9 @@
     <ClInclude Include="cppcheckexecutorseh.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="cppcheckexecutorsig.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="filelister.h">
       <Filter>Header Files</Filter>
     </ClInclude>
@@ -47,6 +50,9 @@
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="cppcheckexecutorseh.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="cppcheckexecutorsig.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
     <ClCompile Include="filelister.cpp">
